Frameworks: Add per app controls for LP keyguard notifications (1/2)
Nice done by google but the UX is a problem especially for ppl who are using a lot apps and just want
to see from important apps the notifications on the lockscreen.

This commit adds the ability to

- enable/disable per app the keyguard notification at all
- enable/disable per app ongoing notifications on the keyguard

We handle this over the app policy conf file like the other per app notification options.

Change-Id: I3e4641c3413ca0a8920b1680ca5f01f6eb4d0225
Signed-off-by: OwnDroid <askowndroid@gmail.com>

Conflicts:
	core/java/android/app/INotificationManager.aidl
	core/java/android/app/NotificationManager.java
	services/core/java/com/android/server/notification/NotificationManagerService.java
	services/core/java/com/android/server/notification/RankingConfig.java
	services/core/java/com/android/server/notification/RankingHelper.java